Created Hangman.cpp
Hangman is a guessing game - a great game for improving your vocabulary and spelling.
The word to guess is represented by a row of dashes representing each letter of the word. The Player is given some chances to guess the correct word. If the Player can guess the correct word within the limited number of chances player wins else the player loses the game.